@@ -1,25 +1,40 @@
 name: lz_flutter
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-#  retrofit: ^2.0.0
-#  dio: 4.0.0
-#  #toast提示
-#  fluttertoast: ^3.1.3
-#  #时间格式
-#  date_format: ^1.0.9
-#  #数据库
-#  floor: ^0.18.0
-#  synchronized: '>=1.5.1 <4.0.0'
-#  #缓存
-#  shared_preferences: '>=0.4.1 <3.0.0'
-#  #路径
-#  path_provider: '1.6.11'
-#  json_annotation: ^3.0.0
-#  #设备信息
-#  device_info: ^0.4.0+1
-#  package_info: ^0.4.0+2
+  #toast提示
+  fluttertoast: ^8.0.7
+  #网络请求
+  retrofit: ^2.0.0
+  dio: 4.0.0
+  #时间格式
+  date_format: ^2.0.2
+  #数据库
+  floor: ^1.1.0
+  synchronized: '>=1.5.1 <4.0.0'
+  #缓存
+  shared_preferences: 2.0.6
+  #路径
+  path_provider: ^2.0.2
+  json_annotation: ^4.0.1
+  #设备信息
+  device_info: ^2.0.2
+  #包信息
+  package_info: ^2.0.2
+  #指纹登录
+  local_auth: ^1.1.6
+  #第三方登录
+  firebase_auth: ^1.3.0
+  google_sign_in: ^5.0.4
+  sign_in_with_apple: ^3.0.0
+  #推送
+  flutter_local_notifications: ^6.0.0
+  flutter_apns: ^1.5.0
+  firebase_messaging: ^9.0.0
+  #依赖注入
+  injectable: ^1.4.1
+  get_it: ^7.1.3
