@@ -14,11 +14,11 @@ import 'package:synchronized/synchronized.dart';
 
 /// SharedPreferences Util.
 class SpUtil {
-  static SpUtil _singleton;
-  static SharedPreferences _prefs;
+  static SpUtil? _singleton;
+  static SharedPreferences? _prefs;
   static Lock _lock = Lock();
 
-  static Future<SpUtil> getInstance() async {
+  static Future<SpUtil?> getInstance() async {
     if (_singleton == null) {
       await _lock.synchronized(() async {
         if (_singleton == null) {
@@ -40,146 +40,142 @@ class SpUtil {
   }
 
   /// put object.
-  static Future<bool> putObject(String key, Object value) {
-    if (_prefs == null) return null;
-    return _prefs.setString(key, value == null ? "" : json.encode(value));
+  static Future<bool>? putObject(String key, Object value) {
+    return _prefs?.setString(key, json.encode(value));
   }
 
   /// get obj.
-  static T getObj<T>(String key, T f(Map v), {T defValue}) {
-    Map map = getObject(key);
+  static T? getObj<T>(String key, T f(Map v), {T? defValue}) {
+    Map? map = getObject(key);
     return map == null ? defValue : f(map);
   }
 
   /// get object.
-  static Map getObject(String key) {
-    if (_prefs == null) return null;
-    String _data = _prefs.getString(key);
+  static Map? getObject(String key) {
+    String? _data = _prefs?.getString(key);
     return (_data == null || _data.isEmpty) ? null : json.decode(_data);
   }
 
   /// put object list.
-  static Future<bool> putObjectList(String key, List<Object> list) {
-    if (_prefs == null) return null;
-    List<String> _dataList = list?.map((value) {
+  static Future<bool>? putObjectList(String key, List<Object> list) {
+    List<String>? _dataList = list.map((value) {
       return json.encode(value);
-    })?.toList();
-    return _prefs.setStringList(key, _dataList);
+    }).toList();
+    return _prefs?.setStringList(key, _dataList);
   }
 
   /// get obj list.
-  static List<T> getObjList<T>(String key, T f(Map v),
-      {List<T> defValue = const []}) {
-    List<Map> dataList = getObjectList(key);
-    List<T> list = dataList?.map((value) {
+  static List<T>? getObjList<T>(String key, T f(Map v),
+      {List<T>? defValue = const []}) {
+    List<Map>? dataList = getObjectList(key);
+    List<T>? list = dataList?.map((value) {
       return f(value);
-    })?.toList();
+    }).toList();
     return list ?? defValue;
   }
 
   /// get object list.
-  static List<Map> getObjectList(String key) {
-    if (_prefs == null) return null;
-    List<String> dataLis = _prefs.getStringList(key);
+  static List<Map>? getObjectList(String key) {
+    List<String>? dataLis = _prefs?.getStringList(key);
     return dataLis?.map((value) {
       Map _dataMap = json.decode(value);
       return _dataMap;
-    })?.toList();
+    }).toList();
   }
 
   /// get string.
-  static String getString(String key, {String defValue = ''}) {
-    if (_prefs == null) return defValue;
-    return _prefs.getString(key) ?? defValue;
+  static String? getString(String key, {String? defValue = ''}) {
+    return _prefs?.getString(key) ?? defValue;
   }
 
   /// put string.
-  static Future<bool> putString(String key, String value) {
-    if (_prefs == null) return null;
-    return _prefs.setString(key, value);
+  static Future<bool>? putString(String key, String value) {
+    return _prefs?.setString(key, value);
   }
 
   /// get bool.
-  static bool getBool(String key, {bool defValue = false}) {
-    if (_prefs == null) return defValue;
-    return _prefs.getBool(key) ?? defValue;
+  static bool? getBool(String key, {bool? defValue = false}) {
+    return _prefs?.getBool(key) ?? defValue;
   }
 
   /// put bool.
-  static Future<bool> putBool(String key, bool value) {
-    if (_prefs == null) return null;
-    return _prefs.setBool(key, value);
+  static Future<bool>? putBool(String key, bool value) {
+    return _prefs?.setBool(key, value);
   }
 
   /// get int.
-  static int getInt(String key, {int defValue = 0}) {
-    if (_prefs == null) return defValue;
-    return _prefs.getInt(key) ?? defValue;
+  static int? getInt(String key, {int? defValue = 0}) {
+    return _prefs?.getInt(key) ?? defValue;
   }
 
   /// put int.
-  static Future<bool> putInt(String key, int value) {
-    if (_prefs == null) return null;
-    return _prefs.setInt(key, value);
+  static Future<bool>? putInt(String key, int value) {
+    return _prefs?.setInt(key, value);
   }
 
   /// get double.
-  static double getDouble(String key, {double defValue = 0.0}) {
-    if (_prefs == null) return defValue;
-    return _prefs.getDouble(key) ?? defValue;
+  static double? getDouble(String key, {double? defValue = 0.0}) {
+    return _prefs?.getDouble(key) ?? defValue;
   }
 
   /// put double.
-  static Future<bool> putDouble(String key, double value) {
-    if (_prefs == null) return null;
-    return _prefs.setDouble(key, value);
+  static Future<bool>? putDouble(String key, double value) {
+    return _prefs?.setDouble(key, value);
   }
 
   /// get string list.
-  static List<String> getStringList(String key,
-      {List<String> defValue = const []}) {
-    if (_prefs == null) return defValue;
-    return _prefs.getStringList(key) ?? defValue;
+  static List<String>? getStringList(String key,
+      {List<String>? defValue = const []}) {
+    return _prefs?.getStringList(key) ?? defValue;
   }
 
   /// put string list.
-  static Future<bool> putStringList(String key, List<String> value) {
-    if (_prefs == null) return null;
-    return _prefs.setStringList(key, value);
+  static Future<bool>? putStringList(String key, List<String> value) {
+    return _prefs?.setStringList(key, value);
   }
 
   /// get dynamic.
-  static dynamic getDynamic(String key, {Object defValue}) {
-    if (_prefs == null) return defValue;
-    return _prefs.get(key) ?? defValue;
+  static dynamic getDynamic(String key, {Object? defValue}) {
+    return _prefs?.get(key) ?? defValue;
   }
 
   /// have key.
-  static bool haveKey(String key) {
-    if (_prefs == null) return null;
-    return _prefs.getKeys().contains(key);
+  static bool? haveKey(String key) {
+    return _prefs?.getKeys().contains(key);
+  }
+
+  /// contains Key.
+  static bool? containsKey(String key) {
+    return _prefs?.containsKey(key);
   }
 
   /// get keys.
-  static Set<String> getKeys() {
-    if (_prefs == null) return null;
-    return _prefs.getKeys();
+  static Set<String>? getKeys() {
+    return _prefs?.getKeys();
   }
 
   /// remove.
-  static Future<bool> remove(String key) {
-    if (_prefs == null) return null;
-    return _prefs.remove(key);
+  static Future<bool>? remove(String key) {
+    return _prefs?.remove(key);
   }
 
   /// clear.
-  static Future<bool> clear() {
-    if (_prefs == null) return null;
-    return _prefs.clear();
+  static Future<bool>? clear() {
+    return _prefs?.clear();
+  }
+
+  /// Fetches the latest values from the host platform.
+  static Future<void>? reload() {
+    return _prefs?.reload();
   }
 
   ///Sp is initialized.
   static bool isInitialized() {
     return _prefs != null;
+  }
+
+  /// get Sp.
+  static SharedPreferences? getSp() {
+    return _prefs;
   }
 }
