@@ -9,7 +9,7 @@ extension ExtString on String {
       LocalManager.getInstance().local(context,this) ?? this;
 
   List<String> resStringList(BuildContext context) =>
-      LocalManager.getInstance().stringList(context,this) ?? null;
+      LocalManager.getInstance().stringList(context,this) ?? [this];
 
 
 }
